Move MSVC environment setup to driver/tool.cpp [NFC]
